display block prevents external CSS
The application I added these UI components into is markdown driven and has it's own styles for use in the documentation. The CSS for these code blocks was interfering with the styling for this component, inserting a `display: grid` property that simulated adding new lines `/n` to the system message. After running a console.log() of chat messages I determined this was not the case and found the bug in CSS.